pyproject.toml
[project]
name = "meshtastic-mcp"
version = "0.1.0"
description = "MCP server for interacting with Meshtastic devices over USB/serial."
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.0",
    "meshtastic>=2.5.0",
    "pyserial>=3.5",
    "pypubsub>=4.0.3",
]

[project.scripts]
meshtastic-mcp = "meshtastic_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["meshtastic_mcp"]
